Improve the env vars management in the graphql-server
