+++ litmus-portal/graphql-server/pkg/cluster/cluster_jwt.go
@@ -3,12 +3,9 @@ package cluster
 import (
 	"errors"
 	"fmt"
-	"os"
-
 	"github.com/golang-jwt/jwt"
+	"github.com/litmuschaos/litmus/litmus-portal/graphql-server/utils"
 )
-
-var secret = os.Getenv("JWT_SECRET")
 
 // ClusterCreateJWT generates jwt used in cluster registration
 func ClusterCreateJWT(id string) (string, error) {
@@ -16,7 +13,7 @@ func ClusterCreateJWT(id string) (string, error) {
 	claims["cluster_id"] = id
 	token := jwt.NewWithClaims(jwt.SigningMethodHS256, claims)
 
-	tokenString, err := token.SignedString([]byte(secret))
+	tokenString, err := token.SignedString([]byte(utils.Config.JwtSecret))
 	if err != nil {
 		return "", err
 	}
@@ -30,7 +27,7 @@ func ClusterValidateJWT(token string) (string, error) {
 		if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
 			return nil, fmt.Errorf("unexpected signing method: %v", token.Header["alg"])
 		}
-		return []byte(secret), nil
+		return []byte(utils.Config.JwtSecret), nil
 	})
 
 	if err != nil {
