--- litmus-portal/graphql-server/utils/constants.go
@@ -1,8 +1,0 @@
-package utils
-
-const (
-	DefaultLitmusAuthRPCPort  = ":3030"
-	DefaultLitmusAuthEndpoint = "localhost"
-	DefaultHTTPPort           = "8080"
-	DefaultRPCPort            = "8000"
-)
